confirm deleting a user in admin dashboard

+++ SkillSwapFrontEnd/src/app/components/admin/admin-dashboard/admin-dashboard.component.ts
@@ -2,7 +2,7 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from '../../../services/user.service';
 import { User } from '../../../models/user.model';
-import {NgForOf} from "@angular/common";
+import {NgForOf, NgIf} from "@angular/common";
 import {RouterLink} from '@angular/router';
 
 @Component({
@@ -10,7 +10,7 @@ import {RouterLink} from '@angular/router';
   templateUrl: './admin-dashboard.component.html',
   imports: [
     NgForOf,
-    RouterLink,
+    RouterLink
   ],
   styleUrls: ['./admin-dashboard.component.css']
 })
@@ -25,13 +25,16 @@ export class AdminDashboardComponent implements OnInit {
 
   getAllUsers(): void {
     this.userService.getAllUsers().subscribe(users => {
-      this.users = users.filter(user=>user.role !== "ADMIN");
+      this.users = users.filter(user =>user.role !== "ADMIN");
     });
   }
 
   deleteUser(id: number): void {
-    this.userService.deleteUser(id).subscribe(() => {
-      this.users = this.users.filter(user => user.id !== id);
-    });
+    const confirmer = confirm("are you sure you want to delete this user ?");
+    if (confirmer){
+      this.userService.deleteUser(id).subscribe(() => {
+        this.users = this.users.filter(user => user.id !== id);
+      });
+    }
   }
 }
